First committal: Uploading project files so far
So far I have created interfaces and the skeleton of basic implementations for the broadcaster and subscriber classes.
The strategy pattern was less interesting so I initially skipped it. This is only a couple hours of work and is my first java (and git use) in years.
Speaking of which, I manually uploaded this push because I couldn't manage it by command line. Something about not being able to read from the remote repository. I'll try debugging that another day.